requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
pyspark>=3.5.0
streamlit>=1.30.0
plotly>=5.18.0
python-pptx>=0.6.23
python-docx>=1.1.0
